Chodec uz ma actualize

+++ src/Enviroment/MovingObject.java
@@ -7,16 +7,16 @@ public class MovingObject {
 	public double speed = 0;
 	public double maxSlowdown = 0;
 	
-	public int posX = 0;
-	public int posY = 0;
+	public double posX = 0;
+	public double posY = 0;
 	
 	public double lastRedrawed = 0;
 
-	public int returnX(){
+	public double returnX(){
 		return posX;
 	}
 	
-	public int returnY(){
+	public double returnY(){
 		return posY;
 	}
